% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addAdmin.R
\name{addAdmin}
\alias{addAdmin}
\title{Add admin information to data.frame}
\usage{
addAdmin(x)
}
\arguments{
\item{x}{A data.frame treated with plantR::formatLoc()}
}
\value{
The same data.frame, with added columns "loc.correct.admin", "country.correct", "stateProvince.correct", "municipality.correct", "locality.correct" and "source.loc", taken from the result of plantR::getAdmin()
}
\description{
Add admin information to data.frame
}
\author{
Mali Oz C Salles
}
