% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{combineSearchStrings}
\alias{combineSearchStrings}
\title{Combine search strings}
\usage{
combineSearchStrings(x)
}
\arguments{
\item{x}{An array of patterns}
}
\description{
Combine strings into a single string that correspond to searching for all strings
}
