% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTaxonId.R
\name{completeScientificName}
\alias{completeScientificName}
\title{completeScientificName}
\usage{
completeScientificName(
  total,
  rm.miss = FALSE,
  na.values = c("indeterminado", "indeterminada", "ndeterminado", "indet", "indet.",
    "sp.", "plantae")
)
}
\arguments{
\item{total}{A data.frame containing identification information with columns "scientificName" and other columns}

\item{rm.miss}{Remove data with no identification info? Defaults to FALSE}

\item{na.values}{Values that if present in scientificName will make that name into NA}
}
\description{
completeScientificName
}
