% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidateCase.R
\name{consolidateCase}
\alias{consolidateCase}
\title{Consolidate Case}
\usage{
consolidateCase(x, ns = goodNames)
}
\arguments{
\item{x}{A data.frame}

\item{ns}{A list of names for reference}
}
\description{
Unifies columns that have identical names except for case
}
