% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_sheets.R
\name{download_sheets}
\alias{download_sheets}
\title{Download a GoogleSheets file}
\usage{
download_sheets(link, sheets = NULL, n = 1, ...)
}
\arguments{
\item{link}{the url of the GoogleSheets file}

\item{sheets}{the names of sheets to download}

\item{n}{the number of sheets to download (use only if sheets is unavailable)}
}
\value{
list of data.frame
}
\description{
This function downloads n sheets from a GoogleSheets file as a list
}
\examples{
data <- download_sheets(DICTIONARIES_URL, 5)
}
