% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixDatum.R
\name{fixDatum}
\alias{fixDatum}
\title{fixDatum}
\usage{
fixDatum(x, convert.to = "EPSG:4674", na = "EPSG:4674")
}
\arguments{
\item{x}{A non-empty data.frame of sf points}

\item{convert.to}{Desired datum. Values will be passed to function sf::st_crs(). Defaults to SIRGAS 2000 (EPSG:4674).}

\item{na}{What datum do you assume NA or unkown values to be. Defaults to SIRGAS 2000 (EPSG:4674).}
}
\description{
Unifies all datum for a data.frame of sf points. Datum information must be in a column "geodeticDatum"
}
