% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixLocation.R
\name{fixLocation}
\alias{fixLocation}
\title{Fix Location}
\usage{
fixLocation(dt, selectedCountry = "Brazil")
}
\arguments{
\item{dt}{A data.frame treated with plantR::formatDwc() and plantR::formatOcc()}

\item{selectedCountry}{The name of a country, usually "Brazil".}
}
\description{
A more elaborate version of plantR::formatLoc, focused on fixing locations within Brazil, especially in São Paulo.
}
