% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatData.R
\name{formatGBIF}
\alias{formatGBIF}
\title{Specific formatting for GBIF data}
\usage{
formatGBIF(x)
}
\arguments{
\item{x}{A data.frame (output from readx)}
}
\description{
Specific formatting for GBIF data
}
