% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatData.R
\name{formatJabot}
\alias{formatJabot}
\title{Specific formatting for Jabot data}
\usage{
formatJabot(x)
}
\arguments{
\item{x}{A data.frame (output from readJabot)}
}
\value{
A data.frame formatted with formatDwc
}
\description{
Specific formatting for Jabot data
}
