% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatData.R
\name{formatOccurrence}
\alias{formatOccurrence}
\title{Generic formatting for darwinCore data}
\usage{
formatOccurrence(x)
}
\arguments{
\item{x}{A data.frame (output from readOccurrence)}
}
\value{
A data.frame formatted with formatDwc
}
\description{
Generic formatting for darwinCore data
}
