% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatData.R
\name{formatReflora}
\alias{formatReflora}
\title{Specific formatting for Reflora data}
\usage{
formatReflora(x)
}
\arguments{
\item{x}{A data.frame (output from readReflora)}
}
\value{
A data.frame formatted with formatDwc
}
\description{
Specific formatting for Reflora data
}
