% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatData.R
\name{formatSpLink}
\alias{formatSpLink}
\title{Specific formatting for SpLink data}
\usage{
formatSpLink(x)
}
\arguments{
\item{x}{A data.frame (output from readSpLink)}
}
\value{
A data.frame formatted with formatDwc
}
\description{
Specific formatting for SpLink data
}
