% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species_and_genus.R
\name{get_species_and_genus}
\alias{get_species_and_genus}
\title{Get species and genus}
\usage{
get_species_and_genus(x)
}
\description{
plantR doesn't return these by default so I have to do it myself
}
