% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTaxonId.R
\name{getTaxonId}
\alias{getTaxonId}
\title{getTaxonId}
\usage{
getTaxonId(total, complete = TRUE, ...)
}
\arguments{
\item{total}{A data.frame containing identification information with columns "scientificName", "scientificNameAuthorship", "genus"}

\item{complete}{Run all possible strategies? If false, will run a default formatTax. Defaults to TRUE}
}
\description{
Try to get taxon ID using many different strategies
}
