% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isolateAuthorship.R
\name{isolateAuthorship}
\alias{isolateAuthorship}
\title{Isolate Authorship}
\usage{
isolateAuthorship(
  x,
  save.original.as = "verbatimScientificName",
  overwrite.authorship = TRUE,
  save.original.authorship.as = "verbatimScientificNameAuthorship",
  tax.name = "scientificName",
  tax.author = "scientificNameAuthorship",
  ...
)
}
\arguments{
\item{x}{A data frame containing scientificName information}

\item{save.original.as}{Name for the column where original scientific name will be saved}

\item{overwrite.authorship}{Logical. If true, will try to extract from all taxon names. If false, records that already have a value in the scientificNameAuthorship column will be ignored. Defaults to TRUE.}

\item{save.original.authorship.as}{Name for the column where original scientificNameAuthorship will be saved, if overwrite.authorship is TRUE}

\item{tax.name}{Name for the column with scientificName information in x}

\item{tax.author}{Name for the column with scientificNameAuthorship information in x}

\item{...}{Aditional parameters to be passed to fixAuthors()}
}
\value{
The x data.frame with corrected information in tax.name and tax.author columns, plus new columns save.original.as and save.original.authorship.as. If these columns are already present in x, they will be overwritten.
}
\description{
Extract authorship information from scientific name
}
\details{
This function uses plantR::fixAuthors() to extract authorship information
}
