% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isolateProblemCases.R
\name{isolateProblemCases}
\alias{isolateProblemCases}
\title{Isolate Problem Cases}
\usage{
isolateProblemCases(
  x,
  FUN,
  breaks = 10,
  parallel = FALSE,
  no_cores = detectCores() - 1,
  ...
)
}
\description{
Isolate Problem Cases
}
