% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_summary.R
\name{make_summary}
\alias{make_summary}
\title{Make summary}
\usage{
make_summary(
  data,
  column,
  levels,
  labels = levels,
  UC = sapply(data, function(x) {
     if ("UC" \%in\% names(x)) 
         x$UC[1]
    
    else if ("Nome_UC" \%in\% names(x)) 
         x$Nome_UC[1]
 })
)
}
\arguments{
\item{data}{A list of data.frames}

\item{column}{The name of the columns to be summarized}

\item{levels}{Possible values in the column (used in factor())}

\item{labels}{Rename the levels, as in factor()}

\item{UC}{Array of names (same length as data)}
}
\description{
This function factorizes a column in each data frame to make a coherent summary to be tabulated
}
