% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{nameSearch}
\alias{nameSearch}
\title{Search for several different entries}
\usage{
nameSearch(ucs, corpus, LT, degrees = "Ouro")
}
\arguments{
\item{ucs}{A data.frame containing names}
}
\description{
Search for several different entries
}
