% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseReflora.R
\encoding{UTF-8}
\name{parseReflora}
\alias{parseReflora}
\title{Converts data from Reflora to DarwinCore}
\usage{
parseReflora(data)
}
\arguments{
\item{data}{Data.table downloaded from Reflora}
}
\description{
Converts data from Reflora to DarwinCore
}
\author{
Mali Oz Salles and Pablo Melo
}
\keyword{internal}
