% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantRWorkflow.R
\name{plantRWorkflow_part1}
\alias{plantRWorkflow_part1}
\title{plantRWorkflow_part1}
\usage{
plantRWorkflow_part1(x)
}
\arguments{
\item{x}{A data.frame containing occurrence data.}
}
\description{
This function is part of the plantR workflow, and is used to format and validate occurrence and location data.
}
