% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatData.R
\name{readGBIF}
\alias{readGBIF}
\title{readGBIF}
\usage{
readGBIF(file, ...)
}
\arguments{
\item{file}{Filename to pass to readOccurrence or readData}
}
\value{
data.frame
}
\description{
Save some options to read properly from GBIF
}
