% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatData.R
\name{readOccurrence}
\alias{readOccurrence}
\title{Read darwin core data}
\usage{
readOccurrence(file, ...)
}
\description{
Basic funtion to read data
}
\examples{

x <- readOccurrence("data-input/Occurrences/OtherSources/example.csv")
}
