% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatData.R
\name{readReflora}
\alias{readReflora}
\title{readeReflora}
\usage{
readReflora(file, ...)
}
\arguments{
\item{file}{Filename to pass to readOccurrence}
}
\value{
data.frame
}
\description{
Save some options to read properly from Reflora
}
