% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatData.R
\name{readSpLink}
\alias{readSpLink}
\title{readSpLink}
\usage{
readSpLink(file, ...)
}
\arguments{
\item{file}{Filename to pass to readOccurrence}
}
\value{
data.frame
}
\description{
Save some options to read properly from SpeciesLink
}
