% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTaxonId.R
\name{removeRepeatedAuthorship}
\alias{removeRepeatedAuthorship}
\title{Remove repeated authorship inside scientific name}
\usage{
removeRepeatedAuthorship(x)
}
\arguments{
\item{x}{A data.frame with columns scientificName and scientificNameAuthorship}
}
\description{
Remove repeated authorship inside scientific name
}
