% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_list_file.R
\name{reverseListFormatting}
\alias{reverseListFormatting}
\title{Return a checklist file to a DWC format}
\usage{
reverseListFormatting(x)
}
\arguments{
\item{x}{A data.frame}
}
\value{
A data.frame formatted with formatDwc
}
\description{
Return a checklist file to a DWC format
}
