% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{searchLoc}
\alias{searchLoc}
\title{Search for a string in location fields of a data.frame}
\usage{
searchLoc(pattern, corpus, fieldA = "locality", fieldB = "municipality")
}
\arguments{
\item{pattern}{Pattern to lookup}

\item{corpus}{A data.frame with locality information}
}
\description{
Search for a string in location fields of a data.frame
}
