% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_utils.R
\name{selectDesiredFields}
\alias{selectDesiredFields}
\title{Select fields}
\usage{
selectDesiredFields(x, desiredFields = desired_fields)
}
\arguments{
\item{x}{A data frame}

\item{desiredFields}{A list of names, ideally contained in names(x)}
}
\description{
A function to reduce the number of columns during data treatement
}
