% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizeNames.R
\name{shorten_uc_name}
\alias{shorten_uc_name}
\title{Shorten UC Name}
\usage{
shorten_uc_name(x)
}
\arguments{
\item{x}{UC names}
}
\description{
Convert conservation unit name to short form (abbreviated)
}
\details{
This function assumes that the names have already been stantardized. It will not standardize names not correct mistakes, and will not work properly if the names are misspelled.
}
