% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizeNames.R
\name{slug}
\alias{slug}
\title{Create slug from UC Name}
\usage{
slug(x)
}
\arguments{
\item{x}{UC names}
}
\description{
Convert conservation unit name into slug for file naming
}
\details{
This function assumes that the names have already been stantardized. It will not standardize names not correct mistakes, and will not work properly if the names are misspelled.
}
