% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizeNames.R
\name{standardize_uc_name}
\alias{standardize_uc_name}
\title{Standardize UC Name}
\usage{
standardize_uc_name(x)
}
\arguments{
\item{x}{UC names}
}
\description{
Convert conservation unit name to long form and correct mistakes
}
\details{
In addition to correcting common mistakes and coverting abbreviations to long form, this function also removes connectors between
}
