% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantRWorkflow.R
\name{subsetToProvince}
\alias{subsetToProvince}
\title{Subset to Province}
\usage{
subsetToProvince(x)
}
\arguments{
\item{x}{A data.frame containing occurrence data.}
}
\description{
This function subsets a data.frame to a specific country and state/province, and removes records that have no location information.
}
