% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tryAgain.R
\name{tryAgain}
\alias{tryAgain}
\title{Try Again}
\usage{
tryAgain(
  x,
  condition,
  FUN,
  success_condition = function(x) !condition(x),
  add_cols = FALSE,
  label = NULL,
  ...
)
}
\arguments{
\item{x}{A data.frame}

\item{condition}{A function that checks which lines must be tried again}

\item{FUN}{A function that takes a data frame, alters it and returns it}

\item{add_cols}{Logical. If TRUE, cols added by the function FUN will be added to the result. Defaults to FALSE.}

\item{...}{Aditional parameters to pass to FUN}
}
\description{
Try to run some function on a data.frame again, but only on rows that have passed some condition, then merge back to the original data.frame, but only the rows that don't pass condition.
}
\details{
You may pass aditional parameters to the function FUN. This function assumes that FUN preserves the order of the rows of x, and that condition() may be applied to both x and FUN(x). FUN may change the values of any columns of x, but only on rows that pass condition before FUN and do not pass condition after FUN.
}
\examples{

x <- data.frame(a = 1:100, b = 100:1)
y <- tryAgain(x, function(x) {x$a > x$b}, function(x) {x$b <- x$b*2; x})
}
