% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{consolidateNamesTable}
\alias{consolidateNamesTable}
\title{Consolidate names table}
\usage{
consolidateNamesTable(
  checkedLocations = utils::read.csv("results/locations/checkedLocations.csv"),
  extraNames = c()
)
}
\arguments{
\item{checkedLocations}{A table containing alternative names information}
}
\description{
Consolidate names table
}
