% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadIPTResource.R
\name{downloadIPTResource}
\alias{downloadIPTResource}
\title{Download IPT Resource}
\usage{
downloadIPTResource(
  resource,
  baseUrl = "https://ipt.jbrj.gov.br/jabot/resource?r=",
  directory = here::here("data-tmp", "JABOT"),
  filename = paste0(resource, ".zip")
)
}
\arguments{
\item{resource}{Name of the resource (ie, collection code)}

\item{baseUrl}{Start of the URL, lacking the resouce name}

\item{directory}{Destination folder}

\item{filename}{Destination file}
}
\description{
Download IPT Resource
}
\author{
Guilherme Gritz

Mali Oz Salles
}
