% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatData.R
\name{formatGBIF}
\alias{formatGBIF}
\title{Specific formatting for GBIF data}
\usage{
formatGBIF(gbif)
}
\arguments{
\item{gbif}{A data.frame (output from readGBIF)}
}
\description{
Specific formatting for GBIF data
}
